+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # dev
 
+* added spring boot template
 * replaced livereload by BrowserSync
 * input prompts are always required fields
 * all templates can be customized right after choosing the template